refactor(residual_tensor): Moved 'residual_tensor' into generation utils.
