+++ hottbox/metrics/__init__.py
@@ -3,4 +3,4 @@ This module includes score functions and performance metrics for the
 computational methods defined in ``hottbox.algorithms``
 """
 
-from .base import mse, rmse, mape, residual_rel_error
+from .decomposition import mse, rmse, mape, residual_rel_error
