+++ hottbox/utils/generation/basic.py
@@ -1,6 +1,6 @@
-import numpy as np
-from hottbox.core.structures import Tensor
 import itertools
+import numpy as np
+from hottbox.core.structures import Tensor, BaseTensorTD
 
 
 def _predefined_distr(distr, shape):
@@ -188,3 +188,29 @@ def super_symmetric_tensor(shape, tensor=None):
     for i, _ in enumerate(inds):
         data = data + np.transpose(tensor.data, tuple(inds[i, :]))
     return Tensor(array=data)
+
+
+def residual_tensor(tensor_orig, tensor_approx):
+    """ Residual tensor
+
+    Parameters
+    ----------
+    tensor_orig : Tensor
+    tensor_approx : {Tensor, TensorCPD, TensorTKD, TensorTT}
+
+    Returns
+    -------
+    residual : Tensor
+    """
+    if not isinstance(tensor_orig, Tensor):
+        raise TypeError("Unknown data type of original tensor.\n"
+                        "The available type for `tensor_A` is `Tensor`")
+    # TODO: make use of direct subtraction of tensors
+    if isinstance(tensor_approx, Tensor):
+        residual = Tensor(tensor_orig.data - tensor_approx.data)
+    elif isinstance(tensor_approx, BaseTensorTD):
+        residual = Tensor(tensor_orig.data - tensor_approx.reconstruct().data)
+    else:
+        raise TypeError("Unknown data type of the approximation tensor!\n"
+                        "The available types for `tensor_B` are `Tensor`,  `TensorCPD`,  `TensorTKD`,  `TensorTT`")
+    return residual
